requirements.txt
python-dotenv
sarvamai
openai